app/proj/requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
black==24.3.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dependency-injector==4.41.0
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-storages==1.14.2
djangorestframework==3.15.1
et-xmlfile==1.1.0
Flask==3.0.2
fonttools==4.50.0
gunicorn==21.2.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mizani==0.11.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
paramiko==3.4.0
pathspec==0.12.1
patsy==0.5.6
pendulum==3.0.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.20.0
plotnine==0.13.3
psycopg2==2.9.9
py4j==0.10.9.7
pyarrow==15.0.2
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.1.2
PyPrind==2.11.3
pysftp==0.2.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
retrying==1.3.4
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
statsmodels==0.14.1
tenacity==8.2.3
time-machine==2.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
whitenoise==6.6.0
zipp==3.18.1
